speech/requirements.txt
tensorflow==2.15.0.post1
tensorboard==2.15.0
transformers==4.52.3
jiwer==3.13.0
librosa==0.10.1
soundfile==0.12.1
hydra-core==1.3.2
tqdm==4.67.1